Adding the Tracklisting Component Information To verify your work locally, run the following command in a terminal: npm run test:part5. The output will let you know which tasks are passing and failing.
To verify your work on Pluralsight, commit and push your changes to GitHub and click the Check My Work button.

Creating The Product Tracklisting Component

Watch video

Our ProductDescriptionComponent is now displaying the album name, artist name, and release date with dynamic data that’s returned from a service class that’s making HTTP requests and getting a JSON response, and all of the data is typed so that our code is better documented and our IDE can catch common errors related to types, but the tracklist data is still hard-coded in the ProductPageComponent. We have a Track interface now, so we should use that to model the data in a new ProductTracklistingComponent.

Start by creating that component with ng generate component product-tracklisting.

Moving The Product Tracklisting HTML

Watch video

Before we edit the TypeScript file for this component, open up product-page.component.html and move the div with a class of row tracklisting and all of its children into product-tracklisting.component.html.

Replace the now empty space in product-page.component.html with the opening and closing tags <app-product-tracklisting></app-product-tracklisting>.

Importing The Album Interface Into The ProductTracklistingComponent

Watch video

Now, in product-tracklisting.component.ts we’ve got to repeat those same steps we’ve done in a few components now - import the interface and service, create a class property to store data returned by the service, inject the service in the constructor, and make an HTTP request with the service to get data into the class property so it’s available in the component’s HTML template.

Let’s start by importing the Album interface from ../album.

Importing The ProductService Into The ProductTracklistingComponent

Watch video

Still in product-tracklisting.component.ts, import the ProductService class from ../product.service.

Creating An albumInfo Property In The ProductTracklistingComponent

Watch video

Still in product-tracklisting.component.ts, create a class property named albumInfo with a type of Album.

Injecting A Private ProductService Object In ProductTracklistingComponent

Watch video

Still in product-tracklisting.component.ts, inject a private instance of the ProductService named _productService in the constructor() function.

Calling getAlbum from Product Tracklisting’s ngOnInit() Method

Watch video

Still in product-tracklisting.component.ts, in the ngOnInit() function, call this._productService.getAlbum(1) and chain a call to the subscribe() method.

For the parameter of that subscribe() method, write response => this.albumInfo = response.